Fix slice creation from system buffers
GetLogonSessions attempts to create a pointer to a slice of LUID objects
by re-interpreting a pointer to the underlying data pointer.

This is invalid. The function will work correctly sometimes because the
first element in the slice header is a pointer to the underlying data.
However, nothing ensures that the length and capacity of the slice are
initialized correctly.

Calling GetLogonSessions can panic with an index out-of-range error if
the length of the slice is interpreted as smaller than the actual number
of elements read from the slice.

Resolve this by declaring a nil slice and initializing all the required
fields correctly.

Also, correct a go vet error in stringFromLSAString (invalid creation of
unsafe.Pointer from uintptr).